Lv4

+++ src/main/java/com/ch3schedulerdevelop/scheduler/service/SchedulerService.java
@@ -21,8 +21,8 @@ public class SchedulerService {
     private final UserRepository userRepository;
 
     @Transactional
-    public CreateSchedulerResponse saveScheduler(CreateSchedulerRequest request){
-        User user = userRepository.findById(request.getUserId()).orElseThrow(
+    public CreateSchedulerResponse saveScheduler(Long sessionUserId, CreateSchedulerRequest request){
+        User user = userRepository.findById(sessionUserId).orElseThrow(
                 () -> new IllegalArgumentException("없는 유저입니다.")
         );
         Scheduler scheduler = new Scheduler(
@@ -35,17 +35,18 @@ public class SchedulerService {
     }
 
     @Transactional(readOnly = true)
-    public List<GetAllSchedulerResponse> getAllSchedulerResponse() {
+    public SchedulerListResponse getAllSchedulerResponse() {
         List<Scheduler> schedulers = schedulerRepository.findAll();
 
-        return schedulers.stream()
+       List<GetAllSchedulerResponse> list = schedulers.stream()
                 .map(GetAllSchedulerResponse::from)
                 .toList();
+        return SchedulerListResponse.of(list);
     }
 
     @Transactional(readOnly = true)
-    public GetAllSchedulerResponse getOneScheduler(Long userId){
-        Scheduler scheduler = schedulerRepository.findById(userId).orElseThrow(
+    public GetAllSchedulerResponse getOneScheduler(Long schedulerId){
+        Scheduler scheduler = schedulerRepository.findById(schedulerId).orElseThrow(
                 () -> new IllegalStateException("존재하지 않는 일정입니다.")
         );
         return GetAllSchedulerResponse.from(scheduler);
@@ -53,10 +54,13 @@ public class SchedulerService {
     }
 
     @Transactional
-    public UpdateSchedulerResponse updateSchedulerResponse(Long userId, UpdateSchedulerRequest request){
-        Scheduler scheduler = schedulerRepository.findById(userId).orElseThrow(
+    public UpdateSchedulerResponse updateSchedulerResponse(Long schedulerId, Long sessionUserId ,UpdateSchedulerRequest request){
+        Scheduler scheduler = schedulerRepository.findById(schedulerId).orElseThrow(
                 () -> new IllegalStateException("존재하지 않는 일정입니다.")
         );
+        if (!scheduler.getUser().getId().equals(sessionUserId)){
+            throw new IllegalStateException("본인의 일정만 수정할 수 있습니다.");
+        }
         scheduler.update(
                 request.getTitle(),
                 request.getContent()
@@ -66,10 +70,13 @@ public class SchedulerService {
     }
 
     @Transactional
-    public void deleteScheduler(Long userId){
-        Scheduler scheduler = schedulerRepository.findById(userId).orElseThrow(
+    public void deleteScheduler(Long schedulerId, Long sessionUserId){
+        Scheduler scheduler = schedulerRepository.findById(schedulerId).orElseThrow(
                 () -> new IllegalStateException("없는 일정입니다.")
         );
+        if (!scheduler.getUser().getId().equals(sessionUserId)){
+            throw new IllegalStateException("본인의 일정만 수정할 수 있습니다.");
+        }
         schedulerRepository.delete(scheduler);
     }
 }
